Fix :bug: where the UI list was not cleared.
Clear the internal row map when the activity manager is stopped. This
fixes a bug where after starting the `ProximityKitManager` a second time
the UI list is not updated. This happened because the inner row map was
not cleared, so the beacon was thought to already have been seen, thus
it was not re-added to the view; however the view itself was already
cleared.