Update the app to Proximity Kit Android v0.3.0
Use the newer Google Play service 5.x to allow more recent phones to use
the application.

It has always been possible to register custom beacon formats. This now
adds a documented example of how to do this with the fictional format
byte sequence: "c0decafe". This also shows that a beacon could possible
include additional data fields in the advertisement.

With the new Proximity Kit library, a new configuration setting has been
added: `setMaxRegionsBeforeRollup`. This includes a documented example
of how to utilize this setting.