Fix for Google Play Services 4.2 and 4.3
This adds a few helper methods to fix some errors due to improper
support for Google Play services 4.2. This changes the method call to
use a pending intent and an error dialog; both of which are consistent
APIs across versions since 4.2

This also utilizes better error checking to handle modded phones which
may not have the Google Play Store app and thus cause issues with the
pending intent provided by Google.