Bump reference app to SDK 21 and Google Play 6.5
This bumps the reference application to SDK 21 and Google Play services
6.5. This uses the new 0.4.0 release with improved support for Google
Play services v6.

In order to keep the minimum SDK as 7 we need to tell the build process
to ignore the minimum required by Google Play services. This is done by
using the `tools:overrideLibrary` attribute in the manifest's `uses-sdk`
element.