Prep for Proximity Kit for Android 0.10.0 release
This switches to using an explicit, properly synchronized, singleton
instance of the `ProximityKitManager`. It uses the new
`ProximityKitManager#getInstance` API to create the manager with a
custom configuration. To demonstrate to legacy apps how to update we
continue to use the `ProximityKit.properties` file. However, we move it
to the `/assets` directory so that we can cleanly load during app start.