Update reference app grade for different google play versions
As we’re working with adjustments to the geofencing apis, we’re going
to want to test our code for each of the different versions that we
have an api for, and for that we’re going to need to frequently switch
the google play services version.  In the interest of that, it’s much
better to have the different versions just commented out instead of
having to rewrite them every time.

I’ve also put markers into the Main Activity document showing which
places have to be commented out to avoid throwing conflicts with Sdk
version 19, which doesn’t recognize a few of the methods used.

Also updates to the new aar file which building monsters-android
produces.